Refactor HttpClientFactory integration
Refactor the HttpClientFactory integration to separate the service collection from the HttpClient builders.
Support automatic decompression on HttpClient instances.
